requirements.txt
numpy
opencv-python
h5py
faiss-cpu
sqlite3
matplotlib
psycopg2
dotenv
ultralytics
pyinstaller